thread lock on video download
